Rename files too
